\name{metadata}
\alias{metadata}
\docType{data}
\title{Example data for meta regression testing.}
\description{
Small example data set used in the example on the metatest help page. 
}
\usage{data(metadata)}
\format{
  A data frame with 6 observations on the following 3 variables.
  \describe{
    \item{\code{y}}{The effect sizes.}
    \item{\code{yvar}}{The variances of the effect sizes.}
    \item{\code{x}}{A moderator variable.}
  }
}

\source{
Data were randomly generated.
}
\examples{
data(metadata)
metadata
}
\keyword{datasets}
