\name{metatest-package}
\alias{metatest-package}

\docType{package}

\title{
	metatest fits a metaregression model and the Bartlett and permutation statistics
}

\description{
	\code{metatest} fits meta regression models and computes a number of
	statistics for testing significance of the coefficients.  In meta
	analyses, the assumptions underlying t- and z-tests are frequently not
	met.  In such cases, the permutation test and the Bartlett corrected
	likelihood ratio test provide better results. }

\details{
	\tabular{ll}{
	Package: \tab metatest\cr
	Type: \tab Package\cr
	Version: \tab 1.0\cr
	Date: \tab 2010-09-16\cr
	License: \tab GPL\cr
	LazyLoad: \tab yes\cr
	}
}

\author{
	The code that does the hard work was written by Hilde Huizenga.  Ingmar
	Visser added the interface functions and handled turning code into a
	package.  Maintainer: Ingmar Visser <i.visser@uva.nl> 
}

\references{
	Hilde M. Huizenga, Ingmar Visser & Conor V. Dolan (2010). Hypothesis
	testing in random effects meta-regression, British journal of
	mathematical and statistical psychology, in press.
}

\keyword{ htest }
\keyword{ models }
\keyword{ regression }
